requirements.txt
annotated-types==0.6.0
anyio==4.3.0
fastapi==0.110.1
idna==3.7
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic_core==2.18.1
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
